+++ ProgramFunctions/Template.fnc.php
@@ -14,13 +14,14 @@
  * For program and user.
  *
  * @since 3.6
+ * @since 10.2.3 Get template from last school year (rollover ID)
  *
  * @example GetTemplate( $_REQUEST['modname'], User( 'STAFF_ID' ) )
  *
  * @param string  $modname Specify program name (optional) defaults to current program.
  * @param integer $staff_id User ID (optional), default to logged in User.
  *
- * @return string Empty if no template found, else Default or user template.
+ * @return string Empty if no template found, else user template or last year's or default.
  */
 function GetTemplate( $modname = '', $staff_id = 0 ) {
 
@@ -29,36 +30,43 @@ function GetTemplate( $modname = '', $staff_id = 0 ) {
 		$modname = $_REQUEST['modname'];
 	}
 
+	$rollover_id = 0;
+
 	if ( ! $staff_id )
 	{
 		$staff_id = User( 'STAFF_ID' );
+
+		$rollover_id = User( 'ROLLOVER_ID' );
+	}
+	else
+	{
+		$rollover_id = (int) DBGetOne( "SELECT ROLLOVER_ID
+			FROM staff
+			WHERE STAFF_ID='" . (int) $staff_id . "'" );
 	}
 
 	$staff_id_sql = '';
 
+	if ( $rollover_id )
+	{
+		// @since 10.2.3 Get template from last school year (rollover ID)
+		$staff_id_sql .= ",'" . $rollover_id . "'";
+	}
+
 	if ( $staff_id )
 	{
 		// Fix SQL error when no user in session.
-		$staff_id_sql = ",'" . $staff_id . "'";
+		$staff_id_sql .= ",'" . $staff_id . "'";
 	}
 
-	$template_RET = DBGet( "SELECT TEMPLATE,STAFF_ID
+	$template = DBGetOne( "SELECT TEMPLATE
 		FROM templates
 		WHERE MODNAME='" . $modname . "'
-		AND STAFF_ID IN(0" . $staff_id_sql . ")", [], [ 'STAFF_ID' ] );
+		AND STAFF_ID IN(0" . $staff_id_sql . ")
+		ORDER BY STAFF_ID DESC
+		LIMIT 1" );
 
-	if ( ! $template_RET )
-	{
-		return '';
-	}
-
-	if ( ! isset( $template_RET[ $staff_id ] ) )
-	{
-		// User has no saved template yet, get the default one.
-		$staff_id = 0;
-	}
-
-	return $template_RET[ $staff_id ][1]['TEMPLATE'];
+	return $template ? $template : '';
 }