+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 10.2.3
 - Remove dead link to centresis.org in index.php
 - Fix MySQL error TEXT column used in key specification without a key length in Fields.fnc.php & DisciplineForm.php
 - Add ROLLOVER_ID column to User() in User.fnc.php
+- Get template from last school year (rollover ID) in Template.fnc.php
 
 Changes in 10.2.2
 -----------------